Extracted zone_type into its own source files
