+++ vme/src/CMakeLists.txt
@@ -101,6 +101,7 @@ set(VME_SRCS
         zon_wiz.cpp zon_wiz.h
         zone_reset.cpp zone_reset.h
         zone_reset_cmd.cpp zone_reset_cmd.h
+        zone_type.cpp zone_type.h
         )
 
 # Add all sources to a library file so we can link with other projects and test suites
